+++ endpoints/aposProxyWithExternalFrontend.js
@@ -1,6 +1,8 @@
 import proxy from "../lib/proxy";
+import aposRequest from "../lib/aposRequest";
 
 export async function ALL({ params, request, redirect }) {
-  console.log(`regular proxy: ${request.url}`);
+  console.log(`external frontend proxy: ${request.url}`);
+  request = aposRequest(request);
   return proxy({ params, request, redirect });
 };